ci(uat): gate the single-instance guard by moving it out of the PyObjC import path (#2426)
`wispr_eyes.py --self-test` was never wired into CI, and the recorded reason was
an untested assumption: the module reaches `ApplicationServices` through
`ui_helpers`, so adding it to the RuntimeUAT step would have rested on the hosted
runner having PyObjC, and a wrong guess reddens the required `build-check` for
everybody.

This takes route 3 from the issue, which the issue calls the better outcome
either way: split the part worth gating into `Tests/RuntimeUAT/instance_guard.py`,
which imports `os`, `subprocess` and `sys` and nothing else, and wire ITS
`--self-test` in. The assumption is removed rather than measured. The
UI-driving harness around it stays out of CI, where it could not run headless
regardless.

`running_enviouswispr_instances` and `_require_single_instance` move verbatim,
comments included, and `wispr_eyes.py` re-exports both — `measure_overlay_first_render.py`
calls `_we.running_enviouswispr_instances()` and keeps working unchanged. The
eight guard rows move with them into `run_guard_cases()`, which both entry points
call, so the two can never disagree about what the guard must do.

Evidence:
- `python3 Tests/RuntimeUAT/instance_guard.py --self-test` -> 8/8 passed
- `python3 Tests/RuntimeUAT/wispr_eyes.py --self-test` -> 28/28 passed, the same
  total as origin/main before this change
- import isolation proved by copying `instance_guard.py` alone into an empty
  directory with no sibling on the path: 8/8 passed. Static audit of its imports:
  os, subprocess, sys, types — all stdlib.
- mutation control 3/3 CAUGHT, restore byte-identical, green after:
  drop the own-pid exclusion -> 1 of 8 FAILED; substring instead of the exact
  `.app/Contents/MacOS/EnviousWispr` suffix -> 1 of 8 FAILED; `len(found) != 1`
  weakened to `< 1` -> 3 of 8 FAILED
- `.claude/scripts/check-cited-symbols.py` -> PASSED, 12 symbols, 4 files

The step's own comment asserted "Neither module imports Quartz, so both run
here", which stops being true of the SET the moment a third line is added. It is
replaced by the membership rule it was standing in for, and by why the new module
satisfies it.

Self-audit-grep: grep -rn "running_enviouswispr_instances\|_require_single_instance\|run_guard_cases" --include='*.py' --include='*.md' --include='*.sh' --include='*.yml' Tests/ Sources/ docs/ .github/ scripts/
Self-audit-owner: Tests/RuntimeUAT/instance_guard.py
